requirements.txt
alembic==1.0.5
aniso8601==4.0.1
asn1crypto==0.24.0
astroid==2.0.4
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
Click==7.0
coverage==4.5.3
cryptography==2.3.1
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jwt==0.5.4
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.13
typed-ast==1.1.0
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
